+++ CHANGELOG.md
@@ -6,9 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.0.6] / 2022-10-31
 ### Features
-- Request Test onlyRead
+- Add Tests Project
 ### Tests
 - Add `ricaun.NUnit.Tests`
+- Add `SampleTest.Tests`
 
 ## [1.0.5] / 2022-09-20
 ### Features